api-gateway/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
aiosqlite==0.20.0
SQLAlchemy[asyncio]==2.0.31
pydantic==2.8.0
